requirements.txt
aiohttp==3.7.4
aioboto3==9.0.0